+++ uilayer-core/src/main/java/net/avh4/framework/uilayer/scene/GraphicsOperations.java
@@ -11,7 +11,7 @@ public interface GraphicsOperations {
 
     void drawLine(int startX, int startY, int stopX, int stopY, int argbColor);
 
-    void drawOval(int leftX, int topY, int width, int height, int argbColor);
+    void drawOval(double leftX, double topY, double width, double height, int argbColor);
 
     void drawImage(String image, int destLeftX, int destTopY, int destRightX, int destBottomY,
                    int sourceLeftX, int sourceTopY, int sourceRightX, int sourceBottomY);
