+++ uilayer-core/src/test/java/net/avh4/framework/uilayer/test/TestGraphicsOperations.java
@@ -48,7 +48,7 @@ public class TestGraphicsOperations implements GraphicsOperations {
     }
 
     @Override
-    public void drawOval(int leftX, int topY, int width, int height, int argbColor) {
+    public void drawOval(double leftX, double topY, double width, double height, int argbColor) {
         sb.append("Oval: ");
         sb.append(leftX).append(", ");
         sb.append(topY).append(", ");
